+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## [3.0.0]
 
-- Upgrade to Spring 3.1.0 and CXF 4.0.1
+- Upgrade to Spring 3.1.0, Java 17 and CXF 4.0.1
 - CXF handles the `X-Forwarded-*` headers during WSDL generation
+- `.http` test file
 
 ## [2.0.1]